Added logic to car controller so only owners of the cars can edit/delete them

+++ app/controllers/cars_controller.rb
@@ -10,29 +10,45 @@ class CarsController < ApplicationController
   end
 
   post '/cars' do
+    #make sure someone is logged in
+    #make sure car created is valid
     cars = Car.create(params[:cars])
     redirect '/account'
   end
 
-  get '/cars/:id' do
+  get '/cars/:id' do 
+    
     @car = Car.find(params[:id])
     erb :'/cars/show'
   end
 
   get '/cars/:id/edit' do
-    @car = Car.find(params[:id])
-    erb :'/cars/edit'
+    @car_edit = Car.find(params[:id])
+    if logged_in? && current_user.id == @car_edit.user_id
+      erb:'/cars/edit'
+    else
+      erb:'/cars/error'
+    end
+    
   end
 
   patch '/cars/:id' do
-    car = Car.find(params[:id])
-    car.update(params[:cars])
-    redirect '/cars/#{car.id}'
+    @car_edit = Car.find(params[:id])
+    if logged_in? && current_user.id == @car_edit.user_id
+        @car_edit.update(params[:cars])
+        redirect '/account'
+    else
+      erb:'/cars/error'
+    end
   end
 
   delete '/cars/:id' do
     car = Car.find(params[:id])
-    car.destroy
-    redirect '/account'
+    if logged_in? && current_user.id == car.user_id
+      car.destroy
+      redirect '/account'
+    else
+      erb:'/cars/error'
+    end
   end
 end
